Split A11y into SkipLink and VisuallyHidden components
