+++ packages/components/src/SkipLink/index.ts
@@ -1,0 +1,1 @@
+export * from "./SkipLink"
