+++ packages/components/src/VisuallyHidden/index.ts
@@ -1,2 +1,1 @@
-export * from "./SkipLink"
 export * from "./VisuallyHidden"
