+++ packages/components/src/index.ts
@@ -2,4 +2,5 @@ export * from "./KaizenProvider"
 export * from "./Filter"
 export * from "./FilterButton"
 export * from "./FilterSelect"
-export * from "./A11y"
+export * from "./SkipLink"
+export * from "./VisuallyHidden"
